+++ docs/about/whats_new.rst
@@ -16,6 +16,11 @@ v1.8.4 (???)
 - Follow up fix to (:issue:`1047`) to round scale to nearest integer if very close.
 - Add support for 3D Datasets. (:pull:`1099`)
 - Fix to enable searching by ``geopolygon`` with the type ``MultiPolygon``.  (:issue:`1149`, :pull:`1150`)
+- Added new "license" and "description" properties to `DatasetType` to enable easier access to product information. (:pull:`1143`, :pull:`1144`)
+- Revised the ``Datacube.list_products`` function to produce a simpler and more useful product list table (:pull:`1145`)
+- Add new ``dataset_predicate`` param to ``dc.load`` and ``dc.find_datasets`` for more flexible temporal filtering (e.g. loading data for non-contiguous time ranges such as specific months or seasons over multiple years). (:pull:`1148`, :pull:`1156`)
+- Fix to ``GroupBy`` to ensure output output axes are correctly labelled when sorting observations using ``sort_key`` (:pull:`1157`)
+- ``GroupBy`` is now its own class to allow easier custom grouping and sorting of data (:pull:`1157`)
 
 .. _`notebook examples`: https://github.com/GeoscienceAustralia/dea-notebooks/